third_party/WebKit/LayoutTests/platform/mac-mac10.9/external/wpt/IndexedDB/nested-cloning-small-expected.txt

This is a testharness.js-based test.
PASS small typed array 
PASS blob 
PASS blob with small typed array 
PASS blob with small typed array with key generator 
PASS blob array 
PASS blob array with key generator 
PASS array of blobs and small typed arrays 
FAIL array of blobs and small typed arrays with key generator assert_equals: Expected success event, but got error event instead expected "success" but got "error"
Harness: the test ran to completion.
